backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
google-genai
yt-dlp>=2023.7.6
python-dotenv>=1.0.0 